requirements.txt
torch==2.6.0
torchvision==0.21.0
numpy==1.26.4
hydra-core>=1.3.2
iopath>=0.1.10
pillow>=9.4.0
tqdm
opencv-python==4.7.0.72
submitit
pandas==2.2.2
scikit-image
fvcore
tensorboard
matplotlib
tensordict
onnx

tensorrt==10.9.0.34  --extra-index-url https://pypi.nvidia.com